Add files via upload
The setup corresponding to this script is a multiple busbar PCB (typically 2) with a sensor mounted over each busbar. The physical parameter through the busbar when it is in measurement mode needs to be estimated. The output of the sensor is used to calculate the parameter through inverse matrix multiplication.